Move supervisor-api.ts to device-api/index.ts
