+++ src/supervisor.ts
@@ -6,7 +6,7 @@ import { intialiseContractRequirements } from './lib/contracts';
 import { normaliseLegacyDatabase } from './lib/legacy';
 import * as osRelease from './lib/os-release';
 import * as logger from './logger';
-import SupervisorAPI from './supervisor-api';
+import SupervisorAPI from './device-api';
 import log from './lib/supervisor-console';
 import version = require('./lib/supervisor-version');
 import * as avahi from './lib/avahi';
