+++ test/legacy/41-device-api-v1.spec.ts
@@ -20,7 +20,7 @@ import mockedAPI = require('~/test-lib/mocked-device-api');
 import sampleResponses = require('~/test-data/device-api-responses.json');
 import * as config from '~/src/config';
 import * as logger from '~/src/logger';
-import SupervisorAPI from '~/src/supervisor-api';
+import SupervisorAPI from '~/src/device-api';
 import * as apiBinder from '~/src/api-binder';
 import * as deviceState from '~/src/device-state';
 import * as apiKeys from '~/lib/api-keys';
